backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.1.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0